+++ CHANGELOG.md
@@ -27,6 +27,7 @@ This release drops the compatibility guarantee of [Go 1.18].
 
 ### Added
 
+- The `WithoutTimestamps` option to `go.opentelemetry.io/otel/exporters/stdout/stdoutmetric` to sets all timestamps to zero. (#3828)
 - Support global `MeterProvider` in `go.opentelemetry.io/otel`. (#3818)
   - Use `Meter` for a `metric.Meter` from the global `metric.MeterProvider`.
   - Use `GetMeterProivder` for a global `metric.MeterProvider`.